% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPA function.R
\name{CommunityPowerAnalysis}
\alias{CommunityPowerAnalysis}
\title{Calculate Community Power Values for Each KO in Each Sample}
\usage{
CommunityPowerAnalysis(keggData)
}
\arguments{
\item{keggData}{A data frame or data table with KOs as rows and samples as columns.}
}
\value{
A data frame with each row representing a community-sample-KO triplet and
the corresponding CP value, along with the CP Score which is the sum of
CP values for each sample.
}
\description{
This function calculates the Community Power (CP) values for each KEGG
Orthology (KO) in each sample within a community. CP value is the proportion
of a specific KO in a sample relative to the total count of that KO in the community.
}
\examples{
# Assuming 'keggData' is your data frame or data table
CPAResult <- CommunityPowerAnalysis(keggData)

}
